requirements.txt
Django~=5.1
argus-server[htmx]==2.0.0
django-widget-tweaks
black
coverage
python-dotenv
django-extensions
flake8
djlint
pytest
pytest-django
pytest-env
pytest-docker
responses
PyYAML
sphinx
sphinx-rtd-theme
sphinxcontrib-drawio